Add script for building releases cross platform
With the removal of Travis CI we needed something that would permit
easy, cross platform builds for the hosts-override utility.

I decided the best approach would be to maintain the currently
documented process of attaching binaries to releases in GitHub.
Changing this may lead to breaking Rapid IdP documentation
elsewhere which I don't want to cause.

AAF CI also seems way to heavy for this little utility so I arrived at
a simple bash script that creates the 3 required zip files for
Windows, MacOS and Linux as the documentation highlights are made
available.

Should a new release need to be cut at some point it will need to be
undertaken on a local machine, binaries build by running the script and
uploads being made to GitHub as part of a new release.